requirements.txt

python-dotenv
requests
pandas
sqlalchemy
psycopg2